脚本权限管理

+++ src/main/java/cc/ruok/hammer/engine/EngineFile.java
@@ -1,5 +1,6 @@
 package cc.ruok.hammer.engine;
 
+import cn.hutool.core.io.FileUtil;
 import org.apache.commons.io.FileUtils;
 
 import java.io.File;
@@ -8,13 +9,23 @@ import java.io.IOException;
 public class EngineFile {
 
     private final File file;
+    private final boolean read;
+    private final boolean write;
 
-    public EngineFile(File file) {
+    public EngineFile(File file, Engine engine) {
         this.file = file;
+        String path = FileUtil.getAbsolutePath(engine.getWebSite().getPath()) + File.separator;
+        if (FileUtil.getAbsolutePath(file).startsWith(path)) {
+            read = engine.getWebSite().getPermission("file_read");
+            write = engine.getWebSite().getPermission("file_write");
+        } else {
+            read = engine.getWebSite().getPermission("public_file_read");
+            write = engine.getWebSite().getPermission("public_file_write");
+        }
     }
 
     public String getPath() {
-        return file.getAbsolutePath();
+        return FileUtil.getAbsolutePath(file);
     }
 
     public boolean exists() {
@@ -30,6 +41,7 @@ public class EngineFile {
     }
 
     public String read(String charset) throws EngineException {
+        if (!read) throw new EngineException("no permission.");
         try {
             return FileUtils.readFileToString(file, charset);
         } catch (IOException e) {
@@ -42,6 +54,7 @@ public class EngineFile {
     }
 
     public void write(String str, String charset) throws EngineException {
+        if (!write) throw new EngineException("no permission.");
         try {
             FileUtils.writeStringToFile(file, str, charset);
         } catch (IOException e) {
@@ -54,6 +67,7 @@ public class EngineFile {
     }
 
     public void append(String str, String charset) throws EngineException {
+        if (!write) throw new EngineException("no permission.");
         try {
             FileUtils.writeStringToFile(file, str, charset, true);
         } catch (IOException e) {
@@ -65,7 +79,8 @@ public class EngineFile {
         append(str, "utf8");
     }
 
-    public boolean create() {
+    public boolean create() throws EngineException {
+        if (!write) throw new EngineException("no permission.");
         try {
             return file.createNewFile();
         } catch (IOException e) {
@@ -73,7 +88,8 @@ public class EngineFile {
         }
     }
 
-    public boolean delete() {
+    public boolean delete() throws EngineException {
+        if (!write) throw new EngineException("no permission.");
         return FileUtils.deleteQuietly(file);
     }
